+++ CHANGELOG.md
@@ -19,6 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.0.3] - 2026-03-19
+
+### Added
+
+- CircuitSet indexing with slices
+- Pyquil transpiler support for subspace parameter
+
+### Changed
+
+- save_to_pickle checks for .pkl extension in filename
+
 ## [1.0.2] - 2026-03-17
 
 ### Added
@@ -118,7 +129,8 @@ post-processing
 - Single- and two-qubit benchmarking methods
 - Advanced compilation tools
 
-[Unreleased]: https://github.com/akelhashim/qcal/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/akelhashim/qcal/compare/v1.0.3...HEAD
+[1.0.3]: https://github.com/akelhashim/qcal/releases/tag/v1.0.3
 [1.0.2]: https://github.com/akelhashim/qcal/releases/tag/v1.0.2
 [1.0.1]: https://github.com/akelhashim/qcal/releases/tag/v1.0.1
 [1.0.0]: https://github.com/akelhashim/qcal/releases/tag/v1.0.0